code cleanup for redeem page
