+++ lib/presentation/screens/redeem_reward_screen.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:loyalty_program_frontend/presentation/utils/helpers/size_helper.dart';
+import 'package:loyalty_program_frontend/presentation/widgets/dailogs/confirmation_dailog.dart';
+import 'package:loyalty_program_frontend/presentation/widgets/dailogs/sucess_dailog.dart';
 
 class RedeemRewardScreen extends StatefulWidget {
   const RedeemRewardScreen({super.key});
@@ -20,9 +22,15 @@ class _RedeemRewardScreenState extends State<RedeemRewardScreen> {
       {"credit": 10000, "denomination": 1500},
       {"credit": 15000, "denomination": 2500}
     ],
-    "products":[
-      {"image":"https://xseed-public-static-assets.s3.us-west-2.amazonaws.com/loyalty-program/amazon_card.png"},
-      {"image":"https://images.template.net/wp-content/uploads/2022/06/Coupon-Sizes1.jpg"}
+    "products": [
+      {
+        "image":
+            "https://xseed-public-static-assets.s3.us-west-2.amazonaws.com/loyalty-program/amazon_card.png"
+      },
+      {
+        "image":
+            "https://images.template.net/wp-content/uploads/2022/06/Coupon-Sizes1.jpg"
+      }
     ],
     "currentCredit": 12500,
   };
@@ -184,11 +192,11 @@ class _RedeemRewardScreenState extends State<RedeemRewardScreen> {
                   SizedBox(
                     height: size(constraints, 200),
                     child: Stack(
-                      children: [                        
+                      children: [
                         PageView.builder(
                           itemCount: data['products'].length,
                           scrollDirection: Axis.horizontal,
-                          onPageChanged: (value){
+                          onPageChanged: (value) {
                             setState(() {
                               currentIndex = value;
                             });
@@ -205,16 +213,17 @@ class _RedeemRewardScreenState extends State<RedeemRewardScreen> {
                           },
                         ),
                         Visibility(
-                          visible: currentIndex!=0,
+                          visible: currentIndex != 0,
                           child: Align(
-                            alignment:
-                                kIsWeb ? Alignment.center : Alignment.centerLeft,
+                            alignment: kIsWeb
+                                ? Alignment.center
+                                : Alignment.centerLeft,
                             child: Padding(
-                              padding:
-                                  EdgeInsets.only(right: kIsWeb? size(constraints, 400):0),
+                              padding: EdgeInsets.only(
+                                  right: kIsWeb ? size(constraints, 400) : 0),
                               child: IconButton(
                                   onPressed: () {
-                                     setState(() {
+                                    setState(() {
                                       currentIndex--;
                                     });
                                   },
@@ -226,13 +235,15 @@ class _RedeemRewardScreenState extends State<RedeemRewardScreen> {
                           ),
                         ),
                         Visibility(
-                          visible: data['products'].length>1 && currentIndex+1!=data['products'].length,
+                          visible: data['products'].length > 1 &&
+                              currentIndex + 1 != data['products'].length,
                           child: Align(
-                            alignment:
-                                kIsWeb ? Alignment.center : Alignment.centerRight,
+                            alignment: kIsWeb
+                                ? Alignment.center
+                                : Alignment.centerRight,
                             child: Padding(
-                              padding:
-                                  EdgeInsets.only(left: kIsWeb?size(constraints, 400):0),
+                              padding: EdgeInsets.only(
+                                  left: kIsWeb ? size(constraints, 400) : 0),
                               child: IconButton(
                                   onPressed: () {
                                     setState(() {
@@ -398,166 +409,7 @@ class _RedeemRewardScreenState extends State<RedeemRewardScreen> {
     showDialog(
       context: context,
       builder: (BuildContext context) {
-        return AlertDialog(
-          shape: const RoundedRectangleBorder(
-            borderRadius: BorderRadius.all(Radius.circular(0)),
-          ),
-          content: Container(
-            width: size(constraints, 652),
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              crossAxisAlignment: CrossAxisAlignment.center,
-              children: [
-                Text(
-                  "Coupon Code",
-                  style: TextStyle(
-                      fontSize: size(constraints, 22),
-                      fontWeight: FontWeight.w600),
-                ),
-                SizedBox(
-                  height: size(constraints, 10),
-                ),
-                Image.asset(
-                  'packages/loyalty_program_frontend/assets/images/coupon.png',
-                  width: size(constraints, 64),
-                  height: size(constraints, 64),
-                ),
-                SizedBox(
-                  height: size(constraints, 20),
-                ),
-                Padding(
-                  padding: EdgeInsets.only(
-                      left: kIsWeb ? size(constraints, 28) : 0,
-                      right: kIsWeb ? size(constraints, 10) : 0),
-                  child: RichText(
-                    text: TextSpan(
-                      text: '',
-                      style: DefaultTextStyle.of(context).style,
-                      children: <TextSpan>[
-                        TextSpan(
-                          text: 'Coupon code worth ',
-                          style: TextStyle(
-                            fontWeight: FontWeight.w500,
-                            color: Colors.black,
-                            fontSize: kIsWeb
-                                ? size(constraints, 18)
-                                : size(constraints, 14),
-                          ),
-                        ),
-                        TextSpan(
-                          text: '₹1,000 ',
-                          style: TextStyle(
-                              color: Colors.black,
-                              fontSize: kIsWeb
-                                  ? size(constraints, 18)
-                                  : size(constraints, 14),
-                              fontWeight: FontWeight.bold),
-                        ),
-                        TextSpan(
-                          text: 'will be send to your registered email ID ',
-                          style: TextStyle(
-                            fontWeight: FontWeight.w500,
-                            color: Colors.black,
-                            fontSize: kIsWeb
-                                ? size(constraints, 18)
-                                : size(constraints, 14),
-                          ),
-                        ),
-                        TextSpan(
-                          text: 'ayushgupta@gmail.com ',
-                          style: TextStyle(
-                              color: Colors.black,
-                              fontSize: kIsWeb
-                                  ? size(constraints, 18)
-                                  : size(constraints, 14),
-                              fontWeight: FontWeight.bold),
-                        ),
-                        TextSpan(
-                          text: 'and phone number ',
-                          style: TextStyle(
-                            fontWeight: FontWeight.w500,
-                            color: Colors.black,
-                            fontSize: kIsWeb
-                                ? size(constraints, 18)
-                                : size(constraints, 14),
-                          ),
-                        ),
-                        TextSpan(
-                          text: '+91-9876567833.',
-                          style: TextStyle(
-                              color: Colors.black,
-                              fontSize: kIsWeb
-                                  ? size(constraints, 18)
-                                  : size(constraints, 14),
-                              fontWeight: FontWeight.bold),
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-                SizedBox(
-                  height: size(constraints, 30),
-                ),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    InkWell(
-                      onTap: () {
-                        Navigator.pop(context);
-                      },
-                      child: Container(
-                        height: kIsWeb
-                            ? size(constraints, 44)
-                            : size(constraints, 34),
-                        width: kIsWeb
-                            ? size(constraints, 258)
-                            : size(constraints, 120),
-                        decoration: BoxDecoration(
-                            color: Colors.white,
-                            border: Border.all(
-                                color: const Color(0xffBA181C), width: 1.2),
-                            borderRadius: BorderRadius.circular(2)),
-                        alignment: Alignment.center,
-                        child: Text(
-                          "Cancel",
-                          style: TextStyle(
-                              fontWeight: FontWeight.w500,
-                              color: const Color(0xffBA181C),
-                              fontSize: size(constraints, 16)),
-                        ),
-                      ),
-                    ),
-                    SizedBox(
-                      width: size(constraints, 20),
-                    ),
-                    InkWell(
-                      onTap: () {},
-                      child: Container(
-                        height: kIsWeb
-                            ? size(constraints, 44)
-                            : size(constraints, 34),
-                        width: kIsWeb
-                            ? size(constraints, 258)
-                            : size(constraints, 120),
-                        decoration: BoxDecoration(
-                            color: const Color(0xffBA181C),
-                            borderRadius: BorderRadius.circular(2)),
-                        alignment: Alignment.center,
-                        child: Text(
-                          "Confirm",
-                          style: TextStyle(
-                              fontWeight: FontWeight.w500,
-                              color: Colors.white,
-                              fontSize: size(constraints, 16)),
-                        ),
-                      ),
-                    ),
-                  ],
-                )
-              ],
-            ),
-          ),
-        );
+        return ConfirmationDailogBox(constraints:constraints);
       },
     );
   }
@@ -566,117 +418,7 @@ class _RedeemRewardScreenState extends State<RedeemRewardScreen> {
     showDialog(
       context: context,
       builder: (BuildContext context) {
-        return AlertDialog(
-          shape: const RoundedRectangleBorder(
-            borderRadius: BorderRadius.all(Radius.circular(0)),
-          ),
-          content: Container(
-            width: size(constraints, 652),
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              crossAxisAlignment: CrossAxisAlignment.center,
-              children: [
-                Text(
-                  "Congratulations!",
-                  style: TextStyle(
-                      fontSize: size(constraints, 22),
-                      fontWeight: FontWeight.w600),
-                ),
-                SizedBox(
-                  height: size(constraints, 10),
-                ),
-                Image.asset(
-                  'packages/loyalty_program_frontend/assets/images/offer.png',
-                  width: size(constraints, 64),
-                  height: size(constraints, 64),
-                ),
-                SizedBox(
-                  height: size(constraints, 20),
-                ),
-                Padding(
-                  padding: EdgeInsets.only(
-                      left: kIsWeb ? size(constraints, 28) : 0,
-                      right: kIsWeb ? size(constraints, 10) : 0),
-                  child: RichText(
-                    text: TextSpan(
-                      text: '',
-                      style: DefaultTextStyle.of(context).style,
-                      children: <TextSpan>[
-                        TextSpan(
-                          text:
-                              'Your coupon code has been successfully sent to your registered email ID ',
-                          style: TextStyle(
-                            fontWeight: FontWeight.w500,
-                            color: Colors.black,
-                            fontSize: kIsWeb
-                                ? size(constraints, 18)
-                                : size(constraints, 14),
-                          ),
-                        ),
-                        TextSpan(
-                          text: 'ayushgupta@gmail.com ',
-                          style: TextStyle(
-                              color: Colors.black,
-                              fontSize: kIsWeb
-                                  ? size(constraints, 18)
-                                  : size(constraints, 14),
-                              fontWeight: FontWeight.bold),
-                        ),
-                        TextSpan(
-                          text: 'and phone number ',
-                          style: TextStyle(
-                            fontWeight: FontWeight.w500,
-                            color: Colors.black,
-                            fontSize: kIsWeb
-                                ? size(constraints, 18)
-                                : size(constraints, 14),
-                          ),
-                        ),
-                        TextSpan(
-                          text: '+91 9876567833.',
-                          style: TextStyle(
-                              color: Colors.black,
-                              fontSize: kIsWeb
-                                  ? size(constraints, 18)
-                                  : size(constraints, 14),
-                              fontWeight: FontWeight.bold),
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-                SizedBox(
-                  height: size(constraints, 30),
-                ),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    InkWell(
-                      onTap: () {
-                        Navigator.pop(context);
-                      },
-                      child: Container(
-                        height: size(constraints, 44),
-                        width: size(constraints, 258),
-                        decoration: BoxDecoration(
-                            color: const Color(0xffBA181C),
-                            borderRadius: BorderRadius.circular(2)),
-                        alignment: Alignment.center,
-                        child: Text(
-                          "OK",
-                          style: TextStyle(
-                              fontWeight: FontWeight.w500,
-                              color: Colors.white,
-                              fontSize: size(constraints, 18)),
-                        ),
-                      ),
-                    ),
-                  ],
-                )
-              ],
-            ),
-          ),
-        );
+        return SucessDailogBox(constraints:constraints);
       },
     );
   }
